feat: add mappers, continue SQLDelight integration;

+++ features/entries/impl/src/commonMain/kotlin/com/kindaboii/journal/features/entries/impl/domain/models/Mood.kt
@@ -1,7 +1,7 @@
 package com.kindaboii.journal.features.entries.impl.domain.models
 
 data class Mood(
-    val value: Int,
+    val value: Long,
     val emotions: List<String>,
     val influences: List<String>,
 )
