+++ features/entries/impl/src/commonMain/kotlin/com/kindaboii/journal/features/entries/impl/presentation/components/MoodHeaderBar.kt
@@ -27,7 +27,7 @@ import kotlin.time.Instant
 
 @Composable
 fun MoodHeaderBar(
-    mood: Int,
+    mood: Long,
     emotions: List<String>,
     influences: List<String>,
     time: Instant,
